backend/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
python-dotenv>=1.0
qiskit-ibm-runtime>=0.25
pydantic>=2.7
